Fix `:source-paths` inference

+++ .circleci/integration_test.sh
@@ -23,6 +23,15 @@ fi
 
 grep --silent "== Warnings: 1. Exceptions thrown: 0" output  || exit 1
 
+rm -rf .eastwood*
+if clojure -M:eastwood > output; then
+  echo "Should have failed! Emitted output:"
+  cat output
+  exit 1
+fi
+
+grep --silent "== Warnings: 1. Exceptions thrown: 0" output  || exit 1
+
 cd ../example-green-project || exit 1
 
 rm -rf .eastwood*
@@ -35,6 +44,14 @@ grep --silent "== Warnings: 0. Exceptions thrown: 0" output  || exit 1
 
 rm -rf .eastwood*
 if ! lein with-profile -user run -m eastwood.lint > output; then
+  echo "Should have passed!"
+  exit 1
+fi
+
+grep --silent "== Warnings: 0. Exceptions thrown: 0" output  || exit 1
+
+rm -rf .eastwood*
+if ! clojure -M:eastwood > output; then
   echo "Should have passed!"
   exit 1
 fi
